requirements_test.txt
coverage==4.1
psycopg2>=2.6.2
dj_database_url>=0.4.1
django_extensions
pytest
pytest-django
pytest-mock
pytest-cov
